sub_8071DA4 and getbodysize

+++ src/dungeon_serializer.c
@@ -133,7 +133,7 @@ static Entity* sub_80828E0(s16 id, s16 apparentID, s32 index);
 static Entity* sub_8082A08(s16 id, s16 apparentID, s32 index);
 
 // These externs are from other files
-extern u8 GetBodySize(s16 index);
+extern u8 GetBodySize(s32 index);
 extern void sub_80460F8(DungeonPos *, Item *, u32);
 
 void SaveDungeonState(u8 *buffer, u32 bufLen)
